Fix missed float literal/static_cast

+++ base/tests/test_DataMatrixSP.cpp
@@ -391,7 +391,7 @@ BOOST_AUTO_TEST_CASE(appendToColTest) {
         size_t k = 0;
         for (size_t i = 0; i < rows; i++) {
           for (size_t j = 0; j < cols + 1; j++) {
-            BOOST_CHECK_MESSAGE(m(i, j) == static_cast<double>(k),
+            BOOST_CHECK_MESSAGE(m(i, j) == static_cast<float>(k),
                                 "rows=" << rows << " cols=" << cols << " m(" << i
                                         << "," << j << ")=" << m(i, j) << "!=" << k);
             k++;
